Refactored check_whitespace.c and check_whitespace_test
The test was failing due to a memory leak in the check_whitespace.c due to the "iscleaned" function not freeing the allocations before returning the result. so i added a few lines to deal with this issue.
however this solved the error in the check_whitespace.c but not the errors in the check_whitespace_test. this is because the strip function dynamically allocates but doesn't free the memory cuausing the leak. so i added to lines to every test to deal with this problem as well.